rag_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
qdrant-client==1.6.0
huggingface-hub==0.16.4
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.1.3
python-multipart==0.0.6
httpx==0.25.2
python-jose[cryptography]==3.3.0
rank-bm25==0.2.2
scikit-learn==1.3.0
tiktoken==0.5.1
PyPDF2==3.0.1
python-docx==0.8.11
openai==1.3.0
